+++ Client/app/src/main/java/com/hackathon/optfit/entities/GpsReading.java
@@ -6,7 +6,7 @@ public class GpsReading {
     public Integer id;
     public String timeStamp;
     public Integer userId;
-    public Integer longitude;
-    public Integer latitude;
+    public double longitude;
+    public double latitude;
 
 }
